requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
reportlab>=4.0.0
PyYAML>=6.0
click>=8.1.0
matplotlib>=3.7.0
